requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
bandit==1.8.3
billiard==4.2.1
black==25.1.0
boolean.py==5.0
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.2
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==44.0.2
cssselect2==0.8.0
cyclonedx-python-lib==9.1.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.4
django-autoslug==1.9.9
django-channels==0.7.0
django-cities-light==3.10.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.0
django-htmx==1.23.0
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
duckduckgo_search==8.0.2
Faker==37.1.0
filelock==3.18.0
flake8==7.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.3.15
license-expression==30.4.1
lxml==5.3.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
openai==1.82.0
orjson==3.10.15
oscrypto==1.3.0
packageurl-python==0.16.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pbr==6.1.1
pillow==11.1.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
pipdeptree==2.26.1
platformdirs==4.3.7
pluggy==1.5.0
primp==0.15.0
progressbar2==4.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.9
py-serializable==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydyf==0.11.0
pyflakes==3.3.2
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pyphen==0.17.2
pytest==8.3.5
pytest-django==4.11.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==3.3.0
python-utils==3.9.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
razorpay==1.4.2
redis==6.1.0
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
serializers==0.2.4
setuptools==76.0.0
simplejwt==2.0.1
six==1.17.0
sniffio==1.3.1
social-auth-core==4.6.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
stevedore==5.4.1
svglib==1.5.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
types-requests==2.32.0.20250515
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
yarl==1.20.0
zopfli==0.2.3.post1
zstandard==0.23.0
